ResQRoute/requirements.txt
streamlit
mysql-connector-python
pandas
folium
geopy
plotly
numpy
requests
